requirements.txt
Flask
markdown
openai
bleach
redis
bs4
Flask-Limiter
python-dotenv
Werkzeug
Flask-WTF
flask_sqlalchemy 
psycopg2-binary
sqlalchemy

